+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog for Rapid Photo Downloader
  - Fixed bug #1889770: Fails to run - Could not load the Qt platform plugin
    "xcb"
 
+ - Fixed bug #1891659: Exception when encountering invalid block device
+
 0.9.24 (2020-05-03)
 -------------------